pyproject.toml
[tool.poetry]
# TODO: update the name and description
name = "ionata-django-skeleton"
version = "0.1.0"
description = ""
authors = ["Ionata <webmaster@ionata.com.au>"]

[tool.poetry.dependencies]
python = "^3.6"
celery = {version = "^4.3", extras = ["redis"]}
coreapi = "^2.3"
django = "^2.2"
django-allauth = "^0.39.1"
django-anymail = {version = "^6.1", extras = ["mailgun"]}
django-axes = "^5.0"
django-celery-beat = "^1.5"
django-celery-results = "^1.1"
django-cors-headers = "^3.1"
django-environ = "^0.4.5"
django-extensions = "^2.2"
django-filter = "^2.2"
django-redis = "~=4.10.0"  # does not follow semver
django-rest-auth = "^0.9.5"
django-storages = {version = "^1.7", extras = ["boto3"]}
djangorestframework = "~=3.10.3"  # does not follow semver
djangorestframework-filters = ">=1.0.0.dev0"
djangorestframework-jsonapi = "^2.8"
gunicorn = "^19.9"
psycopg2 = "^2.8"
python-dateutil = "^2.8"
pytz = "*"

[tool.poetry.dev-dependencies]
# debugging
django-debug-toolbar = "^2.0"
ipdb = "^0.11.0"
ipython = "^7.2"
ptpython = "^2.0"
# linting
autopep8 = "^1.4"
black = {version = "^18.3-alpha.0", allows-prereleases = true}
django-stubs = "^1.1"
mypy = "*"
pycodestyle = "^2.4"
pydocstyle = "^3.0"
pylint = "^2.2"
pylint-django = "^2.0"
pyls-isort = "^0.1.1"
pyls-mypy = "^0.1.5"
python-language-server = "^0.21.4"
rope = "^0.11.0"
# testing
pyhamcrest = "^1.9"
# running
werkzeug = "^0.14.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
